Move event validation with other validations
